Restore the name of Telemetry file to keep backward compatible
