+++ AZ3166/src/cores/arduino/az_iot/iothub_client/src/iothub_client_ll.c
@@ -19,7 +19,7 @@
 #include "iothub_client_options.h"
 #include "iothub_client_version.h"
 #include <stdint.h>
-#include "SystemTelemetry.h"
+#include "Telemetry.h"
 
 #ifndef DONT_USE_UPLOADTOBLOB
 #include "iothub_client_ll_uploadtoblob.h"
